requirements.txt
aioserial==1.3.0
pandas==1.3.5
PyInstaller==5.1
PyQt5==5.15.6
pyqtgraph==0.12.3
pyserial==3.5
PyYAML==6.0
Quamash==0.6.1
Pillow
